fix misc. compiler warnings

+++ src/Cafe/HW/Espresso/Recompiler/IML/IMLInstruction.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "Common/precompiled.h"
 using IMLRegID = uint16; // 16 bit ID
 using IMLPhysReg = sint32; // arbitrary value that is up to the architecture backend, usually this will be the register index. A value of -1 is reserved and means not assigned
 
@@ -758,7 +759,7 @@ struct IMLInstruction
 	void make_fpr_compare(IMLReg regA, IMLReg regB, IMLReg regR, IMLCondition cond)
 	{
 		this->type = PPCREC_IML_TYPE_FPR_COMPARE;
-		this->operation = -999;
+		this->operation = (uint8)-999;
 		this->op_fpr_compare.regR = regR;
 		this->op_fpr_compare.regA = regA;
 		this->op_fpr_compare.regB = regB;
@@ -807,7 +808,7 @@ struct IMLInstruction
 	void make_x86_eflags_jcc(IMLCondition cond, bool invertedCondition)
 	{
 		this->type = PPCREC_IML_TYPE_X86_EFLAGS_JCC;
-		this->operation = -999;
+		this->operation = (uint8)-999;
 		this->op_x86_eflags_jcc.cond = cond;
 		this->op_x86_eflags_jcc.invertedCondition = invertedCondition;
 	}
